+++ src/python/pants/backend/awslambda/common/BUILD
@@ -6,7 +6,7 @@ python_library(
   dependencies=[
     'src/python/pants/base:build_root',
     'src/python/pants/build_graph',
-    'src/python/pants/core/rule_utils',
+    'src/python/pants/core/util_rules',
     'src/python/pants/engine:addressable',
     'src/python/pants/engine:console',
     'src/python/pants/engine:fs',
