+++ src/python/pants/backend/awslambda/python/BUILD
@@ -7,7 +7,7 @@ python_library(
     'src/python/pants/backend/python/subsystems',
     'src/python/pants/backend/awslambda/common',
     'src/python/pants/build_graph',
-    'src/python/pants/core/rule_utils',
+    'src/python/pants/core/util_rules',
     'src/python/pants/engine:addressable',
     'src/python/pants/engine:fs',
     'src/python/pants/engine:isolated_process',
