+++ src/python/pants/backend/python/lint/bandit/BUILD
@@ -8,7 +8,7 @@ python_library(
     'src/python/pants/backend/python/subsystems',
     'src/python/pants/backend/python/rules',
     'src/python/pants/core/goals',
-    'src/python/pants/core/rule_utils',
+    'src/python/pants/core/util_rules',
     'src/python/pants/engine:fs',
     'src/python/pants/engine:isolated_process',
     'src/python/pants/engine:rules',
