+++ src/python/pants/backend/python/lint/black/rules.py
@@ -20,8 +20,8 @@ from pants.backend.python.subsystems import python_native_code, subprocess_envir
 from pants.backend.python.subsystems.subprocess_environment import SubprocessEncodingEnvironment
 from pants.core.goals.fmt import FmtConfiguration, FmtConfigurations, FmtResult
 from pants.core.goals.lint import LinterConfigurations, LintResult
-from pants.core.rule_utils import determine_source_files, strip_source_roots
-from pants.core.rule_utils.determine_source_files import (
+from pants.core.util_rules import determine_source_files, strip_source_roots
+from pants.core.util_rules.determine_source_files import (
     AllSourceFilesRequest,
     SourceFiles,
     SpecifiedSourceFilesRequest,
