+++ src/python/pants/backend/python/rules/BUILD
@@ -11,7 +11,7 @@ python_library(
     'src/python/pants/build_graph',
     'src/python/pants/core',
     'src/python/pants/core/goals',
-    'src/python/pants/core/rule_utils',
+    'src/python/pants/core/util_rules',
     'src/python/pants/engine/legacy:graph',
     'src/python/pants/engine:build_files',
     'src/python/pants/engine:fs',
@@ -35,7 +35,7 @@ python_tests(
     'src/python/pants/backend/python/subsystems',
     'src/python/pants/build_graph',
     'src/python/pants/core',
-    'src/python/pants/core/rule_utils',
+    'src/python/pants/core/util_rules',
     'src/python/pants/engine:fs',
     'src/python/pants/engine:isolated_process',
     'src/python/pants/engine:rules',
@@ -62,7 +62,7 @@ python_tests(
     ':rules',
     'src/python/pants/backend/python/subsystems',
     'src/python/pants/core/goals',
-    'src/python/pants/core/rule_utils',
+    'src/python/pants/core/util_rules',
     'src/python/pants/engine:fs',
     'src/python/pants/engine:rules',
     'src/python/pants/engine:selectors',
