+++ src/python/pants/backend/python/rules/pytest_runner_integration_test.py
@@ -26,7 +26,7 @@ from pants.base.specs import FilesystemLiteralSpec, OriginSpec, SingleAddress
 from pants.build_graph.address import Address
 from pants.build_graph.build_file_aliases import BuildFileAliases
 from pants.core.goals.test import Status, TestDebugRequest, TestOptions, TestResult
-from pants.core.rule_utils import determine_source_files, strip_source_roots
+from pants.core.util_rules import determine_source_files, strip_source_roots
 from pants.engine.fs import FileContent
 from pants.engine.interactive_runner import InteractiveRunner
 from pants.engine.rules import RootRule, subsystem_rule
