+++ src/python/pants/backend/python/rules/python_create_binary.py
@@ -22,7 +22,7 @@ from pants.backend.python.rules.targets import (
 from pants.backend.python.rules.targets import PythonPlatforms as PythonPlatformsField
 from pants.backend.python.targets.python_binary import PythonBinary
 from pants.core.goals.binary import BinaryConfiguration, CreatedBinary
-from pants.core.rule_utils.determine_source_files import AllSourceFilesRequest, SourceFiles
+from pants.core.util_rules.determine_source_files import AllSourceFilesRequest, SourceFiles
 from pants.engine.addressable import Addresses
 from pants.engine.rules import UnionRule, rule
 from pants.engine.selectors import Get
