+++ src/python/pants/backend/python/rules/run_setup_py.py
@@ -33,13 +33,13 @@ from pants.backend.python.rules.util import (
 from pants.backend.python.subsystems.subprocess_environment import SubprocessEncodingEnvironment
 from pants.base.specs import AddressSpecs, AscendantAddresses, SingleAddress
 from pants.build_graph.address import Address
-from pants.core.rule_utils.determine_source_files import AllSourceFilesRequest, SourceFiles
-from pants.core.rule_utils.distdir import DistDir
-from pants.core.rule_utils.strip_source_roots import (
+from pants.core.target_types import ResourcesSources
+from pants.core.util_rules.determine_source_files import AllSourceFilesRequest, SourceFiles
+from pants.core.util_rules.distdir import DistDir
+from pants.core.util_rules.strip_source_roots import (
     SourceRootStrippedSources,
     StripSourcesFieldRequest,
 )
-from pants.core.target_types import ResourcesSources
 from pants.engine.addressable import Addresses
 from pants.engine.console import Console
 from pants.engine.fs import (
