+++ src/python/pants/backend/python/rules/util.py
@@ -9,8 +9,8 @@ from typing import Dict, Iterable, List, Optional, Set, Tuple, cast
 from pkg_resources import Requirement
 
 from pants.backend.python.rules.targets import PythonSources
-from pants.core.rule_utils.strip_source_roots import SourceRootStrippedSources
 from pants.core.target_types import ResourcesSources
+from pants.core.util_rules.strip_source_roots import SourceRootStrippedSources
 from pants.engine.fs import FilesContent
 from pants.engine.target import Target
 from pants.python.python_setup import PythonSetup
