+++ src/python/pants/core/goals/BUILD
@@ -8,7 +8,7 @@ python_library(
     'src/python/pants/base:exiter',
     'src/python/pants/base:specs',
     'src/python/pants/build_graph',
-    'src/python/pants/core/rule_utils',
+    'src/python/pants/core/util_rules',
     'src/python/pants/engine:console',
     'src/python/pants/engine:desktop',
     'src/python/pants/engine:fs',
@@ -34,7 +34,7 @@ python_tests(
     'src/python/pants/base:exceptions',
     'src/python/pants/base:specs',
     'src/python/pants/build_graph',
-    'src/python/pants/core/rule_utils',
+    'src/python/pants/core/util_rules',
     'src/python/pants/engine:console',
     'src/python/pants/engine:fs',
     'src/python/pants/engine:interactive_runner',
@@ -56,7 +56,7 @@ python_tests(
     'src/python/pants/backend/python/rules',
     'src/python/pants/backend/python/subsystems',
     'src/python/pants/build_graph',
-    'src/python/pants/core/rule_utils',
+    'src/python/pants/core/util_rules',
     'src/python/pants/engine:fs',
     'src/python/pants/engine:interactive_runner',
     'src/python/pants/engine:rules',
