+++ src/python/pants/core/goals/repl_integration_test.py
@@ -14,7 +14,7 @@ from pants.backend.python.subsystems import python_native_code, subprocess_envir
 from pants.backend.python.targets.python_library import PythonLibrary as PythonLibraryV1
 from pants.build_graph.build_file_aliases import BuildFileAliases
 from pants.core.goals.repl import Repl, run_repl
-from pants.core.rule_utils import strip_source_roots
+from pants.core.util_rules import strip_source_roots
 from pants.engine.fs import FileContent
 from pants.engine.interactive_runner import InteractiveRunner
 from pants.engine.rules import RootRule
