+++ src/python/pants/core/goals/test_test.py
@@ -25,7 +25,7 @@ from pants.core.goals.test import (
     WrappedTestConfiguration,
     run_tests,
 )
-from pants.core.rule_utils.filter_empty_sources import (
+from pants.core.util_rules.filter_empty_sources import (
     ConfigurationsWithSources,
     ConfigurationsWithSourcesRequest,
 )
