+++ src/python/pants/core/util_rules/determine_source_files_test.py
@@ -14,14 +14,14 @@ from pants.base.specs import (
     SingleAddress,
 )
 from pants.build_graph.address import Address
-from pants.core.rule_utils.determine_source_files import (
+from pants.core.target_types import FilesSources
+from pants.core.util_rules.determine_source_files import (
     AllSourceFilesRequest,
     SourceFiles,
     SpecifiedSourceFilesRequest,
 )
-from pants.core.rule_utils.determine_source_files import rules as determine_source_files_rules
-from pants.core.rule_utils.strip_source_roots import rules as strip_source_roots_rules
-from pants.core.target_types import FilesSources
+from pants.core.util_rules.determine_source_files import rules as determine_source_files_rules
+from pants.core.util_rules.strip_source_roots import rules as strip_source_roots_rules
 from pants.engine.selectors import Params
 from pants.engine.target import Sources as SourcesField
 from pants.engine.target import rules as target_rules
