+++ src/python/pants/core/util_rules/filter_empty_sources.py
