+++ src/python/pants/core/util_rules/strip_source_roots.py
