+++ src/python/pants/core/util_rules/strip_source_roots_test.py
@@ -6,13 +6,13 @@ from typing import List, Optional, Type, Union
 import pytest
 
 from pants.build_graph.address import Address
-from pants.core.rule_utils.strip_source_roots import (
+from pants.core.target_types import FilesSources
+from pants.core.util_rules.strip_source_roots import (
     SourceRootStrippedSources,
     StripSnapshotRequest,
     StripSourcesFieldRequest,
 )
-from pants.core.rule_utils.strip_source_roots import rules as strip_source_root_rules
-from pants.core.target_types import FilesSources
+from pants.core.util_rules.strip_source_roots import rules as strip_source_root_rules
 from pants.engine.scheduler import ExecutionError
 from pants.engine.selectors import Params
 from pants.engine.target import Sources as SourcesField
